@@ -18,7 +18,7 @@ class DataImportConfirmDialog extends StatelessWidget {
     return Container(
       width: context.isPhone ? 1.sw : 400,
       decoration: BoxDecoration(
-        color: context.theme.dialogBackgroundColor,
+        color: context.theme.dialogTheme.backgroundColor,
         border: Border.all(width: 0.3, color: context.theme.dividerColor),
         borderRadius: context.isPhone
             ? const BorderRadius.only(
@@ -120,10 +120,10 @@ class DataImportConfirmDialog extends StatelessWidget {
                 Container(
                   padding: const EdgeInsets.all(12),
                   decoration: BoxDecoration(
-                    color: Colors.orange.withOpacity(0.1),
+                    color: Colors.orange.withValues(alpha:0.1),
                     borderRadius: BorderRadius.circular(8),
                     border: Border.all(
-                      color: Colors.orange.withOpacity(0.2),
+                      color: Colors.orange.withValues(alpha:0.2),
                       width: 0.5,
                     ),
                   ),
@@ -141,7 +141,7 @@ class DataImportConfirmDialog extends StatelessWidget {
                           'importWarning'.tr,
                           style: TextStyle(
                             fontSize: 12,
-                            color: context.theme.textTheme.bodySmall?.color?.withOpacity(0.8),
+                            color: context.theme.textTheme.bodySmall?.color?.withValues(alpha:0.8),
                             fontWeight: FontWeight.w500,
                           ),
                         ),
@@ -164,7 +164,7 @@ class DataImportConfirmDialog extends StatelessWidget {
           width: 4,
           height: 4,
           decoration: BoxDecoration(
-            color: context.theme.textTheme.bodyMedium?.color?.withOpacity(0.6),
+            color: context.theme.textTheme.bodyMedium?.color?.withValues(alpha:0.6),
             borderRadius: BorderRadius.circular(2),
           ),
         ),
@@ -173,7 +173,7 @@ class DataImportConfirmDialog extends StatelessWidget {
           text,
           style: TextStyle(
             fontSize: 13,
-            color: context.theme.textTheme.bodyMedium?.color?.withOpacity(0.8),
+            color: context.theme.textTheme.bodyMedium?.color?.withValues(alpha:0.8),
           ),
         ),
       ],
